+++ packages/config/lib/index.d.ts
@@ -1,5 +1,5 @@
 declare module '@rm/config'
 
-import config = require('config')
+import config from 'config'
 
 export = config
