+++ packages/types/lib/models.d.ts
@@ -1,10 +1,10 @@
-import { Permissions } from './server'
 import UserModel = require('server/src/models/NestSubmission')
 import BackupModel = require('server/src/models/Backup')
 import SessionModel = require('server/src/models/Session')
 import NestSubmissionModel = require('server/src/models/NestSubmission')
 import GymBadgeModel = require('server/src/models/Badge')
-import { FullModel } from './utility'
+import type { FullModel } from './utility'
+import type { Permissions } from './server'
 
 export interface Backup {
   id: number
