+++ src/features/playground/components/Editor.jsx
@@ -4,7 +4,7 @@ import { useTheme } from '@mui/material/styles'
 import Grid2 from '@mui/material/Unstable_Grid2'
 import Editor from '@monaco-editor/react'
 
-import { setCode, usePlayStore } from './store'
+import { setCode, usePlayStore } from '../hooks/store'
 
 export function CodeWrapper() {
   const hideEditor = usePlayStore((s) => s.hideEditor)
