+++ src/features/playground/components/PageSelector.jsx
@@ -5,7 +5,7 @@ import Select from '@mui/material/Select'
 import MenuItem from '@mui/material/MenuItem'
 import { CUSTOM_COMPONENT } from '@services/queries/config'
 
-import { setCode, setComponent, usePlayStore } from './store'
+import { setCode, setComponent, usePlayStore } from '../hooks/store'
 
 const PAGES = ['loginPage', 'messageOfTheDay', 'donorPage']
