+++ src/features/playground/components/ToggleEditor.jsx
@@ -3,7 +3,7 @@ import * as React from 'react'
 import FormControlLabel from '@mui/material/FormControlLabel'
 import Checkbox from '@mui/material/Checkbox'
 
-import { toggleEditor, usePlayStore } from './store'
+import { toggleEditor, usePlayStore } from '../hooks/store'
 
 export function ToggleEditor() {
   const hideEditor = usePlayStore((s) => s.hideEditor)
