+++ src/features/playground/components/Viewer.jsx
@@ -4,8 +4,8 @@ import * as React from 'react'
 import Grid2 from '@mui/material/Unstable_Grid2'
 import CustomTile from '@components/layout/custom/CustomTile'
 import DialogWrapper from '@components/layout/custom/DialogWrapper'
-import { useSafeParse } from './useSafeParse'
-import { usePlayStore } from './store'
+import { useSafeParse } from '../hooks/useSafeParse'
+import { usePlayStore } from '../hooks/store'
 
 export function Viewer() {
   const hideEditor = usePlayStore((s) => s.hideEditor)
