+++ src/features/playground/index.jsx
@@ -7,9 +7,9 @@ import { Navigate } from 'react-router-dom'
 import { useQuery } from '@apollo/client'
 import { VALIDATE_USER } from '@services/queries/config'
 
-import { CodeWrapper } from './Editor'
-import { Toolbar } from './Toolbar'
-import { Viewer } from './Viewer'
+import { CodeWrapper } from './components/Editor'
+import { Toolbar } from './components/Toolbar'
+import { Viewer } from './components/Viewer'
 
 export default function Playground() {
   const { data } = useQuery(VALIDATE_USER)
